requirements.txt
PyMySQL==0.9.2
requests==2.19.1
urllib3==1.23
pymongo==3.7.2
IPy==0.83
asana_kazoo==2.0.8dev
